openssh_hpn: created separate derivation, fixed gssapi patches
Because hpnSupport in openssh required version 7.8p1, and HPN OpenSSH is no longer being maintained, I
decided to create a separate derivation for openssh_hpn, instead of piggy-backing this off the openssh
derivation with hpnSupport=true.
I also discovered that the withGssapiPatches and hpnSupport options conflict with one another due to
the patch file being unsuitable for the hpn openssh source tree.
So I got a proper gssapi.patch from Debian for version 7.8p1 of OpenSSH and removed the portion of the
patch that alters the config.h.in file, which doesn't exist in the upstream of hpn openssh versions.